+++ STM32/PROJECT_MINI2/Core/Src/main.c
@@ -19,6 +19,8 @@
 /* Includes ------------------------------------------------------------------*/
 #include "main.h"
 #include "DHT11.h"
+#include "HC_SR04.h"
+#include "SG90.h"
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
 
@@ -42,6 +44,7 @@
 /* Private variables ---------------------------------------------------------*/
 TIM_HandleTypeDef htim1;
 TIM_HandleTypeDef htim2;
+TIM_HandleTypeDef htim3;
 
 /* USER CODE BEGIN PV */
 
@@ -52,6 +55,7 @@ void SystemClock_Config(void);
 static void MX_GPIO_Init(void);
 static void MX_TIM1_Init(void);
 static void MX_TIM2_Init(void);
+static void MX_TIM3_Init(void);
 /* USER CODE BEGIN PFP */
 
 /* USER CODE END PFP */
@@ -60,64 +64,8 @@ static void MX_TIM2_Init(void);
 /* USER CODE BEGIN 0 */
 #define	SPEARKER					GPIO_PIN_1	//PA1
 #define SPEARKER_PORT			GPIOA
-#define TRIG_PIN 					GPIO_PIN_9	//PB9
-#define	TRIG_PORT					GPIOB
-#define ECHO_PIN 					GPIO_PIN_8	//PB8
-#define ECHO_PORT					GPIOB
-
 volatile float test = 0;
-uint32_t test2;
-//API
-void HC_SR04_Init();
-void HC_SR04_Trigger();
-float HC_SR04_Get_Distance();
 
-void HC_SR04_Init(){
-	//Bat dau timer vi HC_SR04 can do chinh xac cao
-	HAL_TIM_Base_Start(&htim2);
-}
-
-void HC_SR04_Trigger(){
-	HAL_GPIO_WritePin(TRIG_PORT, TRIG_PIN, GPIO_PIN_SET);
-	//Datasheet bao kich len 10us ma luoi qua nen de 1ms
-	HAL_Delay(1);
-	//Keo chan Trig xuong thap
-	HAL_GPIO_WritePin(TRIG_PORT, TRIG_PIN, GPIO_PIN_RESET);
-}
-
-float HC_SR04_Get_Distance(){
-	uint32_t start_CNT;
-	uint32_t end_CNT;
-	uint32_t raw_CNT;			//Tick timer t
-	uint32_t timeout = 60; 	//60ms
-	float distance; 
-	//Step 1: 
-	HC_SR04_Trigger();
-	//Wait cho ECHO len cao
-	uint32_t t0 = HAL_GetTick();
-	while(HAL_GPIO_ReadPin(ECHO_PORT, ECHO_PIN)==0){
-		//Neu het 60ms ma ko duoc thi tra ve -1
-		if(HAL_GetTick() - t0 >= timeout){
-			return -1;
-		}
-	}
-	start_CNT = htim2.Instance -> CNT;
-	//Wait cho ECHO xuong thap
-	uint32_t t1 = HAL_GetTick();
-	while(HAL_GPIO_ReadPin(ECHO_PORT, ECHO_PIN)== 1){
-		//Neu het 60ms ma ko duoc thi tra ve -1
-		if(HAL_GetTick() - t1 >= timeout){
-			return -1;
-		}
-	}
-	end_CNT = htim2.Instance -> CNT;
-	//Da luu lai gia tri CNT sau khi ECHO duoc kich
-	raw_CNT = end_CNT - start_CNT;
-	//Chuyen doi 1 lan bang cong thuc eq3: distance = us / 58
-	distance = (float)raw_CNT/58;
-	//Tra ve khoang cach
-	return distance;
-}
 
 /* USER CODE END 0 */
 
@@ -152,9 +100,10 @@ int main(void)
   MX_GPIO_Init();
   MX_TIM1_Init();
   MX_TIM2_Init();
-	
+  MX_TIM3_Init();
   /* USER CODE BEGIN 2 */
 	HC_SR04_Init();
+	servo_init();
   /* USER CODE END 2 */
 
   /* Infinite loop */
@@ -164,14 +113,23 @@ int main(void)
     /* USER CODE END WHILE */
 
     /* USER CODE BEGIN 3 */
-//		DHT_Handle();
-//		HAL_Delay(1000);
-//		HAL_GPIO_WritePin(SPEARKER_PORT, SPEARKER, 1);
-//		HAL_Delay(1000);
-//		HAL_GPIO_WritePin(SPEARKER_PORT, SPEARKER, 0);
-//		HAL_Delay(1000);
 		test = HC_SR04_Get_Distance();
 		HAL_Delay(50);
+		//Neu nhu gan cua, se tit coi 3 lan va den 3 lan
+		if(test <= 10.0){
+			for(uint8_t i =0; i < 3; i++){
+				HAL_GPIO_WritePin(SPEARKER_PORT, SPEARKER, 1);
+				HAL_Delay(250);
+				HAL_GPIO_WritePin(SPEARKER_PORT, SPEARKER, 0);
+				HAL_Delay(250);
+			}
+			//Mo cua 3s
+			servo_change_angle(90);
+			HAL_Delay(3000);
+			//Dong cua
+			servo_change_angle(0);
+		}
+		else HAL_GPIO_WritePin(SPEARKER_PORT, SPEARKER, 0);
   }
   /* USER CODE END 3 */
 }
@@ -303,6 +261,55 @@ static void MX_TIM2_Init(void)
   /* USER CODE BEGIN TIM2_Init 2 */
 
   /* USER CODE END TIM2_Init 2 */
+
+}
+
+/**
+  * @brief TIM3 Initialization Function
+  * @param None
+  * @retval None
+  */
+static void MX_TIM3_Init(void)
+{
+
+  /* USER CODE BEGIN TIM3_Init 0 */
+
+  /* USER CODE END TIM3_Init 0 */
+
+  TIM_MasterConfigTypeDef sMasterConfig = {0};
+  TIM_OC_InitTypeDef sConfigOC = {0};
+
+  /* USER CODE BEGIN TIM3_Init 1 */
+
+  /* USER CODE END TIM3_Init 1 */
+  htim3.Instance = TIM3;
+  htim3.Init.Prescaler = 71;
+  htim3.Init.CounterMode = TIM_COUNTERMODE_UP;
+  htim3.Init.Period = 20000;
+  htim3.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
+  htim3.Init.AutoReloadPreload = TIM_AUTORELOAD_PRELOAD_DISABLE;
+  if (HAL_TIM_PWM_Init(&htim3) != HAL_OK)
+  {
+    Error_Handler();
+  }
+  sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
+  sMasterConfig.MasterSlaveMode = TIM_MASTERSLAVEMODE_DISABLE;
+  if (HAL_TIMEx_MasterConfigSynchronization(&htim3, &sMasterConfig) != HAL_OK)
+  {
+    Error_Handler();
+  }
+  sConfigOC.OCMode = TIM_OCMODE_PWM1;
+  sConfigOC.Pulse = 0;
+  sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
+  sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
+  if (HAL_TIM_PWM_ConfigChannel(&htim3, &sConfigOC, TIM_CHANNEL_1) != HAL_OK)
+  {
+    Error_Handler();
+  }
+  /* USER CODE BEGIN TIM3_Init 2 */
+
+  /* USER CODE END TIM3_Init 2 */
+  HAL_TIM_MspPostInit(&htim3);
 
 }